Fixes to be compatible with the new versions of FakeIt.

+++ include/mockutils/VTUtils.hpp
@@ -65,7 +65,7 @@ namespace fakeit {
 		}
 
         template<typename C>
-        static size_t getVTSize() {
+        static unsigned int getVTSize() {
             struct Derrived : public C {
                 virtual void endOfVt() {
                 }
